Move getElementName to helpers and move timerdata to own files
