+++ companion/src/firmwares/modeldata.cpp
@@ -27,31 +27,6 @@
 #include "helpers.h"
 #include "adjustmentreference.h"
 
-/*
- * TimerData
- */
-
-void TimerData::convert(RadioDataConversionState & cstate)
-{
-  cstate.setComponent(tr("TMR"), 1);
-  cstate.setSubComp(tr("Timer %1").arg(cstate.subCompIdx + 1));
-  mode.convert(cstate);
-}
-
-bool TimerData::isEmpty()
-{
-  return (mode == RawSwitch(SWITCH_TYPE_TIMER_MODE, 0) && name[0] == '\0' && minuteBeep == 0 && countdownBeep == COUNTDOWN_SILENT && val == 0 && persistent == 0 /*&& pvalue == 0*/);
-}
-
-QString TimerData::nameToString(int index) const
-{
-  return RadioData::getElementName(tr("TMR", "as in Timer"), index + 1, name);
-}
-
-/*
- * ModelData
- */
-
 ModelData::ModelData()
 {
   clear();
