+++ companion/src/firmwares/modeldata.h
@@ -33,6 +33,7 @@
 #include "output_data.h"
 #include "sensordata.h"
 #include "telem_data.h"
+#include "timerdata.h"
 
 #include <QtCore>
 
@@ -53,35 +54,6 @@ class RSSIAlarmData {
       this->critical = 42;
       this->disabled = false;
     }
-};
-
-#define TIMER_NAME_LEN 8
-
-class TimerData {
-  Q_DECLARE_TR_FUNCTIONS(TimerData)
-
-  public:
-    enum CountDownMode {
-      COUNTDOWN_SILENT,
-      COUNTDOWN_BEEPS,
-      COUNTDOWN_VOICE,
-      COUNTDOWN_HAPTIC
-    };
-    TimerData() { clear(); }
-    RawSwitch    mode;
-    char         name[TIMER_NAME_LEN+1];
-    bool         minuteBeep;
-    unsigned int countdownBeep;
-    unsigned int val;
-    unsigned int persistent;
-    int          countdownStart;
-    unsigned int direction;
-    int          pvalue;
-    void clear() { memset(reinterpret_cast<void *>(this), 0, sizeof(TimerData)); mode = RawSwitch(SWITCH_TYPE_TIMER_MODE, 0); }
-    void convert(RadioDataConversionState & cstate);
-    bool isEmpty();
-    bool isModeOff() { return mode == RawSwitch(SWITCH_TYPE_TIMER_MODE, 0); }
-    QString nameToString(int index) const;
 };
 
 #define CPN_MAX_SCRIPTS       9
