+++ companion/src/firmwares/sensordata.h
@@ -20,7 +20,7 @@
 
 #pragma once
 
-#include "helpersdata.h"
+#include "datahelpers.h"
 
 #include <QtCore>
